trivial jit harness
We arrange to jit fast-hinted cores with a ~ huk (funks) that do not
have a jet implementation.  We modify the hot state (u3j_core, u3j_harm)
by generating those structures when no driver exists.  For this version,
the jit simply contructs code which calls the interpreter so that we can
test the plumbing in isolation from any optimization.